+++ wifipumpkin3/extensions/proxies.py
@@ -84,24 +84,25 @@ class Proxies(ExtensionUI):
             ):
                 return
             all_plugins = plugin_info_activated["Config"].get_all_childname("plugins")
-            for plugin_name in all_plugins:
-                status_plugin = config_instance.get("plugins", plugin_name, format=bool)
-                output_plugins.append(
-                    [
-                        plugin_name,
-                        setcolor("True", color="green")
-                        if status_plugin
-                        else setcolor("False", color="red"),
-                    ]
+            if all_plugins:
+                for plugin_name in all_plugins:
+                    status_plugin = config_instance.get("plugins", plugin_name, format=bool)
+                    output_plugins.append(
+                        [
+                            plugin_name,
+                            setcolor("True", color="green")
+                            if status_plugin
+                            else setcolor("False", color="red"),
+                        ]
+                    )
+                print(
+                    display_messages(
+                        "{} plugins:".format(plugin_info_activated["Name"]),
+                        info=True,
+                        sublime=True,
+                    )
                 )
-            print(
-                display_messages(
-                    "{} plugins:".format(plugin_info_activated["Name"]),
-                    info=True,
-                    sublime=True,
-                )
-            )
-            display_tabulate(headers_plugins, output_plugins)
+                display_tabulate(headers_plugins, output_plugins)
             
             print(display_messages("Settings:", info=True, sublime=True))
             plugin_settings = plugin_info_activated["Config"].get_all_childname("settings")