+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Added 
 - added: optional to set shared interface internet connection 
 - added: network core command for show information about connections
+- added: phishkin3 proxy for perform custom phishing attack using captive portal 
 
 ## [1.1.5]